requirements.txt
fastapi
uvicorn
python-multipart
google-genai
librosa
soundfile
nest-asyncio
python-dotenv